Remove ActiveModel::Validations from validator
