+++ lib/jsonapi_parameters.rb
@@ -1,6 +1,6 @@
 require 'jsonapi_parameters/parameters'
 require 'jsonapi_parameters/handlers'
-require 'jsonapi_parameters/validator'
+require 'jsonapi_parameters/schema_validator'
 require 'jsonapi_parameters/translator'
 require 'jsonapi_parameters/core_ext'
 require 'jsonapi_parameters/stack_limit'
