+++ lib/jsonapi_parameters/translator.rb
@@ -18,12 +18,12 @@ module JsonApi::Parameters
 
     @jsonapi_unsafe_hash = ensure_naming(params, naming_convention)
 
-    JsonApi::Parameters::Validator.new(@jsonapi_unsafe_hash.deep_dup).validate! if should_prevalidate?
+    JsonApi::Parameters::SchemaValidator.new(@jsonapi_unsafe_hash.deep_dup).validate! if should_prevalidate?
 
     formed_parameters
   rescue StandardError => err
     # Validate the payload and raise errors...
-    JsonApi::Parameters::Validator.new(@jsonapi_unsafe_hash.deep_dup).validate! unless JsonApi::Parameters.suppress_schema_validation_errors
+    JsonApi::Parameters::SchemaValidator.new(@jsonapi_unsafe_hash.deep_dup).validate! unless JsonApi::Parameters.suppress_schema_validation_errors
 
     raise err # ... or if there were none, re-raise initial error
   end
