+++ spec/lib/jsonapi_parameters/validator_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
+describe JsonApi::Parameters::SchemaValidator do # rubocop:disable RSpec/FilePath
   describe 'initializer' do
     it 'ensures @payload has keys deeply stringified' do
       validator = described_class.new(payload: { sample: 'value' })
@@ -26,7 +26,7 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
         it 'raises validation errors' do
           payload = { payload: { sample: 'value' } }
 
-          expect { translator.jsonapify(payload) }.to raise_error(ActiveModel::ValidationError)
+          expect { translator.jsonapify(payload) }.to raise_error(JsonApi::Parameters::SchemaValidator::ValidationError)
         end
 
         it 'does not raise TranslatorError' do
@@ -42,7 +42,7 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
 
           begin
             translator.jsonapify(payload)
-          rescue ActiveModel::ValidationError => _ # rubocop:disable Lint/HandleExceptions
+          rescue JsonApi::Parameters::SchemaValidator::ValidationError => _ # rubocop:disable Lint/HandleExceptions
           end
         end
       end
@@ -55,7 +55,7 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
         it 'does not raise validation errors' do
           payload = { payload: { sample: 'value' } }
 
-          expect { translator.jsonapify(payload) }.not_to raise_error(ActiveModel::ValidationError)
+          expect { translator.jsonapify(payload) }.not_to raise_error(JsonApi::Parameters::SchemaValidator::ValidationError)
         end
 
         it 'still raises any other errors' do
@@ -77,7 +77,7 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
         it 'raises validation errors' do
           payload = { payload: { sample: 'value' } }
 
-          expect { translator.jsonapify(payload) }.to raise_error(ActiveModel::ValidationError)
+          expect { translator.jsonapify(payload) }.to raise_error(JsonApi::Parameters::SchemaValidator::ValidationError)
         end
       end
 
@@ -88,7 +88,7 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
         expect(File).to receive(:read).with(JsonApi::Parameters::SCHEMA_PATH).and_call_original
         expect(JSONSchemer).to receive(:schema).and_call_original
 
-        expect { validator.validate! }.to raise_error(ActiveModel::ValidationError)
+        expect { validator.validate! }.to raise_error(JsonApi::Parameters::SchemaValidator::ValidationError)
       end
     end
 
@@ -98,11 +98,11 @@ describe JsonApi::Parameters::Validator do # rubocop:disable RSpec/FilePath
         payload = { controller: 'examples_controller', action: 'create', commit: 'Sign up' }
         validator = described_class.new(payload)
 
-        expect { validator.validate! }.to raise_error(ActiveModel::ValidationError)
+        expect { validator.validate! }.to raise_error(JsonApi::Parameters::SchemaValidator::ValidationError)
 
         begin
           validator.validate!
-        rescue ActiveModel::ValidationError => err
+        rescue JsonApi::Parameters::SchemaValidator::ValidationError => err
           rails_specific_params.each do |param|
             expect(err.message).not_to include("Payload path '/#{param}'")
           end
